refactor: move NotFoundException to App\Exceptions
