+++ app/Exceptions/NotFoundException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Exceptions;
 
 use RuntimeException;
