+++ app/Services/Themes/ThemeInformationService.php
@@ -3,9 +3,9 @@
 namespace App\Services\Themes;
 
 use App\Data\WpOrg\Themes\ThemeInformationRequest;
+use App\Exceptions\NotFoundException;
 use App\Http\Resources\ThemeResource;
 use App\Models\WpOrg\Theme;
-use App\NotFoundException;
 
 class ThemeInformationService
 {
